requirements.txt
google-adk
python-dotenv>=1.0.0
openai>=1.3.0
google-generativeai>=0.5.0
anthropic>=0.20.0
# pydantic is usually included as a dependency of google-adk or other frameworks
# If you encounter issues, you might need to add a specific version, e.g., pydantic>=2.0.0
